+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -9,6 +9,7 @@ import LogOutIcon from "~/assets/img/svg/log-out.svg?react";
 import PieIcon from "~/assets/img/svg/pie.svg?react";
 import PlusIcon from "~/assets/img/svg/plus.svg?react";
 import PlusOutlinedIcon from "~/assets/img/svg/plus-outlined.svg?react";
+import SearchIcon from "~/assets/img/svg/search.svg?react";
 import { type IconName } from "~/libs/types/types.js";
 
 const iconNameToSvg: Record<
@@ -26,6 +27,7 @@ const iconNameToSvg: Record<
 	pie: PieIcon,
 	plus: PlusIcon,
 	plusOutlined: PlusOutlinedIcon,
+	search: SearchIcon,
 };
 
 export { iconNameToSvg };