+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -9,6 +9,7 @@ type IconName =
 	| "logOut"
 	| "pie"
 	| "plus"
-	| "plusOutlined";
+	| "plusOutlined"
+	| "search";
 
 export { type IconName };